Enable session token support.
This enables session token support by creating a context manager SessionToken which takes a user's session token as an argument and includes it in the header sent to parse.com, allowing queries and saves to ACL protected objects. Thanks to [jimmyho](https://github.com/dgrtwo/ParsePy/issues/74) for getting this started.